Update test checklist, enhance tensor dtype handling, and refactor neural flow components
- Updated the test checklist to reflect the completion status of integration tests.
- Improved tensor handling in `DimensionManager` and `NeuralGeometricFlow` classes by ensuring tensors are created with the correct data type, enhancing compatibility and performance.
- Refactored tensor operations in tests to utilize the appropriate dtype, ensuring consistency across test cases and improving numerical stability.
- Removed unnecessary dependency markers in integration tests for cleaner test management.

These changes enhance the clarity and reliability of the testing framework and improve tensor operations within the neural flow components.

+++ src/core/common/dimensions.py
@@ -365,8 +365,8 @@ class DimensionManager(nn.Module):
         Returns:
             Converted QuantumTensor
         """
-        # Create a new tensor with the same data
-        new_tensor = tensor.clone()
+        # Create a new tensor with the same data and dtype
+        new_tensor = tensor.clone().to(dtype=tensor.dtype)
         # Convert to QuantumTensor
         return QuantumTensor(new_tensor.detach().requires_grad_(tensor.requires_grad))
